resonance structures now account for charges
WARNING: something wrong with procedures for changing molecules, hence some tests fail
