+++ examples/01_toy_minimization/toy_minimization.py
@@ -5,6 +5,7 @@ import numpy as np
 from mosaics import ExtGraphCompound, RandomWalk
 from mosaics.beta_choice import gen_exp_beta_array
 from mosaics.minimized_functions import OrderSlide
+from mosaics.ext_graph_compound import str2ExtGraphCompound
 
 random.seed(1)
 np.random.seed(1)
@@ -44,17 +45,7 @@ global_change_params = {
 }
 
 # All replicas are initialized in methane.
-init_ncharges = [6]
-init_adj_matrix = [[0]]
-
-init_egcs = [
-    ExtGraphCompound(
-        nuclear_charges=init_ncharges,
-        adjacency_matrix=init_adj_matrix,
-        hydrogen_autofill=True,
-    )
-    for _ in betas
-]
+init_egcs = [str2ExtGraphCompound("6#4") for _ in betas]
 
 min_func = OrderSlide(possible_elements=possible_elements)
