+++ mosaics/distributed_random_walk.py
@@ -21,7 +21,7 @@ from .random_walk import (
     default_minfunc_name,
     maintain_sorted_CandidateCompound_list,
 )
-from .valence_treatment import misc_global_variables_current_kwargs, set_misc_global_variables
+from .chem_graph import misc_global_variables_current_kwargs, set_misc_global_variables
 
 
 class SubpopulationPropagationIntermediateResults:
@@ -60,7 +60,9 @@ class SubpopulationPropagationIntermediateResults:
         if self.num_minfunc_saved_steps is None:
             self.minfunc_val_log = None
         else:
-            self.minfunc_val_log = np.empty((self.num_minfunc_saved_steps, len(rw.cur_tps)))
+            self.minfunc_val_log = np.empty(
+                (self.num_minfunc_saved_steps, len(rw.cur_tps))
+            )
 
     def update(self, rw):
         """
@@ -95,11 +97,13 @@ class SubpopulationPropagationIntermediateResults:
                 acc_prob = Metropolis_acceptance_probability(
                     (beta1 - beta2) * (minfunc_1 - minfunc_2)
                 )
-                self.sum_tempering_neighbors_acceptance_probability[exploration_replica_id] += (
-                    acc_prob / self.num_beta_subpopulation_clones
-                )
+                self.sum_tempering_neighbors_acceptance_probability[
+                    exploration_replica_id
+                ] += (acc_prob / self.num_beta_subpopulation_clones)
 
-    def update_worst_accepted_candidates(self, current_trajectory_points, minimized_function_name):
+    def update_worst_accepted_candidates(
+        self, current_trajectory_points, minimized_function_name
+    ):
         if self.worst_accepted_candidates is None:
             self.worst_accepted_candidates = [
                 CandidateCompound(
@@ -258,9 +262,9 @@ class DistributedRandomWalk:
 
         # Correct random walk keyword arguments.
         self.random_walk_kwargs["num_saved_candidates"] = self.num_saved_candidates
-        self.random_walk_kwargs[
-            "saved_candidates_max_difference"
-        ] = self.saved_candidates_max_difference
+        self.random_walk_kwargs["saved_candidates_max_difference"] = (
+            self.saved_candidates_max_difference
+        )
         # Make sure betas and initial conditions are not accidentally defined twice for the random walk.
         for del_key in ["betas", "init_tps", "init_egcs"]:
             if del_key in self.random_walk_kwargs:
@@ -283,10 +287,18 @@ class DistributedRandomWalk:
         )
 
         self.terminated_worker_max_restart_number = terminated_worker_max_restart_number
-        self.terminated_worker_num_extra_rng_calls = terminated_worker_num_extra_rng_calls
-        self.subpopulation_num_extra_rng_calls = np.zeros((self.num_subpopulations,), dtype=int)
-        self.subpopulation_num_attempted_restarts = np.zeros((self.num_subpopulations,), dtype=int)
-        self.subpopulation_propagation_completed = np.zeros((self.num_subpopulations,), dtype=int)
+        self.terminated_worker_num_extra_rng_calls = (
+            terminated_worker_num_extra_rng_calls
+        )
+        self.subpopulation_num_extra_rng_calls = np.zeros(
+            (self.num_subpopulations,), dtype=int
+        )
+        self.subpopulation_num_attempted_restarts = np.zeros(
+            (self.num_subpopulations,), dtype=int
+        )
+        self.subpopulation_propagation_completed = np.zeros(
+            (self.num_subpopulations,), dtype=int
+        )
         self.subpopulation_propagation_results_list = [
             None for _ in range(self.num_subpopulations)
         ]
@@ -312,7 +324,9 @@ class DistributedRandomWalk:
         minfunc_val = init_tp.calc_or_lookup(minfunc_eval_dict)[self.min_function_name]
         return CandidateCompound(init_tp, minfunc_val)
 
-    def initialize_current_trajectory_points(self, init_tps=None, init_egcs=None, init_egc=None):
+    def initialize_current_trajectory_points(
+        self, init_tps=None, init_egcs=None, init_egc=None
+    ):
         if init_tps is None:
             if init_egcs is None:
                 assert init_egc is not None
@@ -328,7 +342,9 @@ class DistributedRandomWalk:
         self.current_trajectory_points = init_tps
 
     def init_cloned_betas(self, beta_input):
-        self.tot_num_clones = self.num_subpopulations * self.num_beta_subpopulation_clones
+        self.tot_num_clones = (
+            self.num_subpopulations * self.num_beta_subpopulation_clones
+        )
         self.original_betas = np.array(beta_input)
         betas = []
         for beta in beta_input:
@@ -401,12 +417,18 @@ class DistributedRandomWalk:
         self.num_valid_crossovers += random_walk_instance.num_valid_crossovers
         self.num_accepted_crossovers += random_walk_instance.num_accepted_crossovers
 
-        self.num_attempted_simple_moves += random_walk_instance.num_attempted_simple_moves
+        self.num_attempted_simple_moves += (
+            random_walk_instance.num_attempted_simple_moves
+        )
         self.num_valid_simple_moves += random_walk_instance.num_valid_simple_moves
         self.num_accepted_simple_moves += random_walk_instance.num_accepted_simple_moves
 
-        self.num_attempted_tempering_swaps += random_walk_instance.num_attempted_tempering_swaps
-        self.num_accepted_tempering_swaps += random_walk_instance.num_accepted_tempering_swaps
+        self.num_attempted_tempering_swaps += (
+            random_walk_instance.num_attempted_tempering_swaps
+        )
+        self.num_accepted_tempering_swaps += (
+            random_walk_instance.num_accepted_tempering_swaps
+        )
 
     def default_init_rng_states(self):
         if self.subpopulation_propagation_seed is None:
@@ -422,7 +444,9 @@ class DistributedRandomWalk:
         for _ in self.original_betas:
             clone_indices = []
             for subpop_id in range(self.num_subpopulations):
-                clone_indices += [subpop_id for _ in range(self.num_beta_subpopulation_clones)]
+                clone_indices += [
+                    subpop_id for _ in range(self.num_beta_subpopulation_clones)
+                ]
             random.shuffle(clone_indices)
             shuffled_indices += clone_indices
         self.subpopulation_indices_list = [[] for _ in range(self.num_subpopulations)]
@@ -456,7 +480,9 @@ class DistributedRandomWalk:
                 all_greedy_indices.append(replica_id)
             else:
                 all_exploration_indices.append(replica_id)
-        subpopulation_greedy_indices_list = self.divide_into_subpopulations(all_greedy_indices)
+        subpopulation_greedy_indices_list = self.divide_into_subpopulations(
+            all_greedy_indices
+        )
         subpopulation_exploration_indices_list = self.divide_into_subpopulations(
             all_exploration_indices
         )
@@ -482,14 +508,18 @@ class DistributedRandomWalk:
             self.update_saved_candidates(candidate)
         if self.track_av_tempering_neighbors_acceptance_probability:
             self.av_tempering_neighbors_acceptance_probability[:] += (
-                cur_intermediate_results.sum_tempering_neighbors_acceptance_probability[:]
+                cur_intermediate_results.sum_tempering_neighbors_acceptance_probability[
+                    :
+                ]
                 / self.num_subpopulations
                 / cur_intermediate_results.nsteps
             )
         if cur_intermediate_results.track_worst_accepted_candidates:
             if self.worst_accepted_candidates is None:
                 self.worst_accepted_candidates = [None for _ in self.betas]
-            self.update_worst_accepted_candidates(subpopulation_indices, cur_intermediate_results)
+            self.update_worst_accepted_candidates(
+                subpopulation_indices, cur_intermediate_results
+            )
         if cur_intermediate_results.minfunc_val_log is not None:
             if self.minfunc_val_log is None:
                 self.minfunc_val_log = np.empty(
@@ -503,17 +533,23 @@ class DistributedRandomWalk:
             ):
                 self.minfunc_val_log[:, replica_index] = replica_minfunc_logs[:]
 
-    def update_worst_accepted_candidates(self, subpopulation_indices, cur_intermediate_results):
+    def update_worst_accepted_candidates(
+        self, subpopulation_indices, cur_intermediate_results
+    ):
         for true_id, worst_cand in zip(
             subpopulation_indices, cur_intermediate_results.worst_accepted_candidates
         ):
             old_candidate = self.worst_accepted_candidates[true_id]
-            if (old_candidate is None) or (old_candidate.func_val < worst_cand.func_val):
+            if (old_candidate is None) or (
+                old_candidate.func_val < worst_cand.func_val
+            ):
                 self.worst_accepted_candidates[true_id] = worst_cand
 
     def completed_subpopulation_sublist(self, full_list):
         sublist = []
-        for completed, element in zip(self.subpopulation_propagation_completed, full_list):
+        for completed, element in zip(
+            self.subpopulation_propagation_completed, full_list
+        ):
             if not completed:
                 sublist.append(element)
         return sublist
@@ -527,7 +563,9 @@ class DistributedRandomWalk:
         """
         all_init_tps = []
         all_betas = []
-        for subpopulation_id, subpopulation_indices in enumerate(self.subpopulation_indices_list):
+        for subpopulation_id, subpopulation_indices in enumerate(
+            self.subpopulation_indices_list
+        ):
             if self.subpopulation_propagation_completed[subpopulation_id]:
                 continue
             # The initial trajectory points and the beta values.
@@ -577,7 +615,8 @@ class DistributedRandomWalk:
             np.logical_not(self.subpopulation_propagation_completed)
         )[0]
         results_iterator = self.executor.map(
-            gen_subpopulation_propagation_result, *self.subpopulation_propagation_inputs()
+            gen_subpopulation_propagation_result,
+            *self.subpopulation_propagation_inputs()
         )
         if self.synchronization_signal_file is not None:
             terminal_run(["rm", "-f", self.synchronization_signal_file])
@@ -585,7 +624,9 @@ class DistributedRandomWalk:
         while True:
             if propagated_subpopulation_id == len(all_propagated_subpopulation_ids):
                 break
-            subpopulation_id = all_propagated_subpopulation_ids[propagated_subpopulation_id]
+            subpopulation_id = all_propagated_subpopulation_ids[
+                propagated_subpopulation_id
+            ]
             try:
                 cur_res = results_iterator.__next__()
                 self.subpopulation_propagation_completed[subpopulation_id] = True
@@ -629,10 +670,16 @@ class DistributedRandomWalk:
                 subpopulation_propagation_latest_logs.append(
                     (subpopulation_indices, random_walk, intermediate_results)
                 )
-            self.update_temporary_data(subpopulation_indices, random_walk, intermediate_results)
+            self.update_temporary_data(
+                subpopulation_indices, random_walk, intermediate_results
+            )
             self.add_to_move_statistics(random_walk)
-            self.subpopulation_random_rng_states[subpopulation_index] = propagation_results[2]
-            self.subpopulation_numpy_rng_states[subpopulation_index] = propagation_results[3]
+            self.subpopulation_random_rng_states[subpopulation_index] = (
+                propagation_results[2]
+            )
+            self.subpopulation_numpy_rng_states[subpopulation_index] = (
+                propagation_results[3]
+            )
             self.update_current_trajectory_points(subpopulation_indices, random_walk)
         if self.save_logs:
             self.propagation_logs.append(subpopulation_propagation_latest_logs)
