+++ mosaics/misc_procedures.py
@@ -2,10 +2,15 @@
 import copy
 import os
 import random
-
+from sortedcontainers import SortedList
 import numpy as np
-
+from itertools import groupby
 from .data import NUCLEAR_CHARGE
+
+
+class InvalidAdjMat(Exception):
+    pass
+
 
 # Related to verbosity.
 
@@ -55,7 +60,9 @@ num_procs_name = "MOSAICS_NUM_PROCS"
 
 
 def default_num_procs(num_procs=None):
-    return checked_environ_val(num_procs_name, expected_answer=num_procs, default_answer=1)
+    return checked_environ_val(
+        num_procs_name, expected_answer=num_procs, default_answer=1
+    )
 
 
 # Sorting-related.
@@ -147,7 +154,9 @@ class weighted_array(list):
                 upper_cutoff = self[remaining_length - 1].rho
                 cut_rho = upper_cutoff * remaining_length + ignored_rhos
                 if cut_rho > (1.0 - remaining_rho):
-                    density_cut = (1.0 - remaining_rho - ignored_rhos) / remaining_length
+                    density_cut = (
+                        1.0 - remaining_rho - ignored_rhos
+                    ) / remaining_length
                     break
                 else:
                     ignored_rhos += upper_cutoff
@@ -211,7 +220,9 @@ class NaturalLogLookup:
         return new_saved_values, old_max_avail_val
 
     def fill_saved_values(self, new_max_avail_val):
-        new_saved_values, old_max_avail_val = self.gen_new_saved_values(new_max_avail_val)
+        new_saved_values, old_max_avail_val = self.gen_new_saved_values(
+            new_max_avail_val
+        )
         for i in range(old_max_avail_val, new_max_avail_val):
             new_saved_values[i] = intlog_no_precalc(i + 1)
         self.saved_values = new_saved_values
@@ -233,7 +244,9 @@ def llenlog(l):
 
 class FactorialLogLookup(NaturalLogLookup):
     def fill_saved_values(self, new_max_avail_val):
-        new_saved_values, old_max_avail_val = self.gen_new_saved_values(new_max_avail_val)
+        new_saved_values, old_max_avail_val = self.gen_new_saved_values(
+            new_max_avail_val
+        )
         if old_max_avail_val == 1:
             current_log = 0.0
         else:
@@ -273,7 +286,9 @@ def random_choice_from_dict(possibilities, choices=None, get_probability_of=None
         corr_prob_choice[choice] = prob
     if get_probability_of is None:
         if len(corr_prob_choice.keys()) == 0:
-            raise Exception("Something is wrong: encountered a molecule that cannot be changed")
+            raise Exception(
+                "Something is wrong: encountered a molecule that cannot be changed"
+            )
         final_choice = random.choices(
             list(corr_prob_choice.keys()), list(corr_prob_choice.values())
         )[0]
@@ -285,7 +300,9 @@ def random_choice_from_dict(possibilities, choices=None, get_probability_of=None
         )
 
 
-def random_choice_from_nested_dict(possibilities, choices=None, get_probability_of=None):
+def random_choice_from_nested_dict(
+    possibilities, choices=None, get_probability_of=None
+):
     continue_nested = True
     cur_possibilities = possibilities
     prob_log = 0.0
@@ -322,3 +339,35 @@ def random_choice_from_nested_dict(possibilities, choices=None, get_probability_
         return final_choice, prob_log
     else:
         return prob_log
+
+
+def list2colors(obj_list):
+    """
+    Color obj_list in a way that each equal obj1 and obj2 were the same color.
+    Used for defining canonical permutation of a graph.
+    """
+    ids_objs = list(enumerate(obj_list))
+    ids_objs.sort(key=lambda x: x[1])
+    num_obj = len(ids_objs)
+    colors = np.zeros(num_obj, dtype=int)
+    cur_color = 0
+    prev_obj = ids_objs[0][1]
+    for i in range(1, num_obj):
+        cur_obj = ids_objs[i][1]
+        if cur_obj != prev_obj:
+            cur_color += 1
+            prev_obj = cur_obj
+        colors[ids_objs[i][0]] = cur_color
+    return colors
+
+
+def merge_unrepeated_sorted_lists(base_list, added_list):
+    for el in added_list:
+        if el not in base_list:
+            base_list.add(el)
+
+
+# from: https://stackoverflow.com/questions/3844801/check-if-all-elements-in-a-list-are-equal
+def all_equal(iterable):
+    g = groupby(iterable)
+    return next(g, True) and not next(g, False)
