+++ mosaics/periodic.py
@@ -93,3 +93,10 @@ p_int = {
     17: 5,
     18: 6,
 }
+
+# Priority list of charged species.
+# Introduced in order charges to C were assigned only if options of fitting valences with adding charges to O and N were exhausted.
+charge_feasibility_list = [[5, 7, 8], [6]]
+max_charge_feasibility = len(charge_feasibility_list)
+# which valences can be
+charged_valences_int = {5: {-1: 4}, 6: {-1: 3}, 7: {1: 4}, 8: {-1: 1}}
