+++ mosaics/utils.py
@@ -42,7 +42,12 @@ def xyz_string(coordinates, elements=None, nuclear_charges=None, extra_string=""
         elements = [str_atom_corr(charge) for charge in nuclear_charges]
     output = str(len(coordinates)) + "\n" + extra_string
     for atom_coords, element in zip(coordinates, elements):
-        output += "\n" + element + " " + " ".join([str(atom_coord) for atom_coord in atom_coords])
+        output += (
+            "\n"
+            + element
+            + " "
+            + " ".join([str(atom_coord) for atom_coord in atom_coords])
+        )
     return output
 
 
@@ -96,7 +101,7 @@ def read_xyz_lines(xyz_lines, additional_attributes=["charge"]):
 
 
 # Break an egc of a disconnected molecules into several egc's corresponding to each fragment.
-def break_into_connected(egc):
+def break_into_connected(egc: ExtGraphCompound):
     output = []
     gc = egc.chemgraph.graph.components()
     mvec = np.array(gc.membership)
@@ -177,13 +182,13 @@ def xbgf2gc(xbgf_file):
 #   Some procedures that often appear in scripts.
 
 
-def egc2xyz_string(egc, extra_string=""):
+def egc2xyz_string(egc: ExtGraphCompound, extra_string=""):
     return xyz_string(
         egc.coordinates, nuclear_charges=egc.nuclear_charges, extra_string=extra_string
     )
 
 
-def write_egc2xyz(egc, xyz_file_name, extra_string=""):
+def write_egc2xyz(egc: ExtGraphCompound, xyz_file_name, extra_string=""):
     write_xyz_file(
         egc.coordinates,
         xyz_file_name,
@@ -236,7 +241,9 @@ def mktmp(directory=False):
     extra_args = ()
     if directory:
         extra_args = ("-d", *extra_args)
-    return subprocess.check_output(["mktemp", *extra_args, "-p", "."], text=True).rstrip("\n")
+    return subprocess.check_output(
+        ["mktemp", *extra_args, "-p", "."], text=True
+    ).rstrip("\n")
 
 
 def mktmpdir():
