+++ tests/03_distributed_random_walk/parallelized_toy_minimization.py
@@ -4,7 +4,7 @@ import sys
 
 import numpy as np
 
-from mosaics import ExtGraphCompound
+from mosaics.ext_graph_compound import str2ExtGraphCompound
 from mosaics.beta_choice import gen_exp_beta_array
 from mosaics.distributed_random_walk import DistributedRandomWalk
 from mosaics.minimized_functions import OrderSlide
@@ -72,17 +72,7 @@ global_step_params = {
 }
 
 # All replicas are initialized in methane.
-init_ncharges = [6]
-init_adj_matrix = [[0]]
-
-init_egcs = [
-    ExtGraphCompound(
-        nuclear_charges=init_ncharges,
-        adjacency_matrix=init_adj_matrix,
-        hydrogen_autofill=True,
-    )
-    for _ in range(num_replicas)
-]
+init_egcs = [str2ExtGraphCompound("6#4") for _ in range(num_replicas)]
 
 min_func = OrderSlide(possible_elements=possible_elements)
 # num_saved_candidates = 40
